Move Version module to Ouroboros.Network.Protocol.VersionNegotiation
