+++ ouroboros-network/src/Ouroboros/Network/Protocol/VersionNegotiation/Version.hs
@@ -4,7 +4,7 @@
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE ConstraintKinds #-}
 
-module Ouroboros.Network.Server.Version
+module Ouroboros.Network.Protocol.VersionNegotiation.Version
   ( Versions (..)
   , Application (..)
   , Version (..)
